WEB/multilanguage_transolator_be/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
google-cloud-translate
google-cloud-storage
boto3
python-dotenv
PyPDF2
python-docx
openpyxl
requests